Honor options when creating loggers, fix comments.
When creating a new logger context, disable the time and change the
output if those options are set. Also, default to TRACE if no level is
set.

When instantiating a subsystem logger, honor the level that's set on it.

Remove the example from the GoDoc string for all our methods, as the
actual examples that will be executed as part of Go test will handle
providing examples of usage that is more likely to stay correct over
time.