Add option to copy os.Stderr at init.
go-plugin's Serve function overwrites os.Stderr with a writer that gets
sent over a gRPC stream. This is great, except currently, Terraform
doesn't actually read from that stream at all. While the ideal fix here
is for Terraform to read from that stream the way it currently does
stderr for the provider process, the current behavior exists, and that
means we need to work around it.

The workaround in this case is to copy os.Stderr at startup, before
Serve gets a chance to overwrite it, and write to that instead of the
overwritten os.Stderr. This mimics the behavior in SDKv2's
helper/logging package.

This isn't ideal long-term behavior, because it's subverting go-plugin's
expectations of what's writing to stderr and when, so it's hidden behind
an Option so it's opt-in. That way we can deprecate and remove the
option at some point in the future, when Terraform has started reading
from the gRPC stream that go-plugin provides and enough practitioners
have upgraded to that version of Terraform.