Refactor into internal package.
An upcoming change is going to introduce a logging sink, which is going
to be a root logger that both our provider and SDK logs need to be
sub-loggers from, so they can inherit its filtering. This means that we
can no longer have the loggers siloed off from each other, or at least
need the sink to be accessible to both.

But that got me thinking: does it make _sense_ to have the loggers in
their own packages, or should the packages be boundaries around intended
audience? The number of "don't use this unless you're building an SDK"
comments in tflog make me sad, and make me wonder if it's better to just
say "tflog is for provider devs, tfsdklog is for SDK devs".

This PR is that. It refactors out the shared stuff between both loggers
into an internal/logging package, then uses that in tflog and tfsdklog.
It also moves things around so that only the functions we expect
provider devs to use are in the tflog package, only the functions we
expect SDK devs to use are in the tfsdklog package, and functions that
are only meant to be used by terraform-plugin-log (for testing, etc.)
are siloed off in the internal/testing package.

This feels much cleaner and less repetitive, and feels like a better
interface for provider devs and SDK devs alike.