Simplified logging sink.
Our logging situation with helper/resource is actually really
complicated, and trying to standardize it is unnecessarily delaying the
implementation of terraform-plugin-log.

This PR simplifies the logging sink, no longer trying to parse
Terraform's logs, and only concerning itself with the provider under
test.

When a new test framework is built or another opportunity for breaking
changes arises, that logging situation can be brought under
terraform-plugin-log and standardized into a proper logging subsystem.